Empty topics cause primary key violation
Unfortunately I had some topics without any postings in my Dizkus forum. This caused primary key constraint violations during importDizkus() and the next topic was not imported.

The reason is that $firstId ist set to the first posting id of the previous topic, if the database query returned no postings for the current topic. To avoid this a topic should be imported only if there is at least one posting for this topic.

This code change is a rather quick workaround. There may be better solutions.